#50 refactor: 코스/장소 dto converter 디렉토리 구조 변경
