+++ src/main/java/com/meetup/hereandnow/course/domain/entity/Course.java
@@ -1,7 +1,7 @@
 package com.meetup.hereandnow.course.domain.entity;
 
 import com.meetup.hereandnow.core.infrastructure.entity.BaseEntity;
-import com.meetup.hereandnow.course.application.service.converter.CourseTagListConverter;
+import com.meetup.hereandnow.course.infrastructure.converter.CourseTagListConverter;
 import com.meetup.hereandnow.member.domain.Member;
 import com.meetup.hereandnow.pin.domain.entity.Pin;
 import jakarta.persistence.*;
