+++ src/test/java/com/meetup/hereandnow/scrap/application/service/CourseScrapServiceTest.java
@@ -6,7 +6,7 @@ import com.meetup.hereandnow.course.infrastructure.repository.CourseRepository;
 import com.meetup.hereandnow.member.domain.Member;
 import com.meetup.hereandnow.scrap.domain.CourseScrap;
 import com.meetup.hereandnow.scrap.dto.response.ScrapResponseDto;
-import com.meetup.hereandnow.scrap.repository.CourseScrapRepository;
+import com.meetup.hereandnow.scrap.infrastructure.repository.CourseScrapRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
