+++ src/test/java/com/meetup/hereandnow/scrap/application/service/PlaceScrapServiceTest.java
@@ -6,7 +6,7 @@ import com.meetup.hereandnow.place.domain.Place;
 import com.meetup.hereandnow.place.infrastructure.repository.PlaceRepository;
 import com.meetup.hereandnow.scrap.domain.PlaceScrap;
 import com.meetup.hereandnow.scrap.dto.response.ScrapResponseDto;
-import com.meetup.hereandnow.scrap.repository.PlaceScrapRepository;
+import com.meetup.hereandnow.scrap.infrastructure.repository.PlaceScrapRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
